IndexMap: add UX improvements
Adds an IndexMap table view option to `settings.yml`. When the option is set to true, the application will append a table row for each IndexMap layer clicked. It will also color layers as "visited" as layers are clicked and added to the table. Clicking a "visited" layer will remove its layer row from the table.

Our end goal will be to add this feature to GBL proper. We'll use GeoData as a means for demonstrating the new functionality for discussions within the GBL community for potential adoption upstream.

Addresses #184